fix watson test
